終盤探索を少し高速化、ヒント処理を評価値順に処理するように変更、ヒントの中断がうまく動いていなかったのを修正

+++ ai_core/bit64.c
@@ -40,7 +40,7 @@ INT32 CountBit(UINT64 bit)
 	///int l_moves = bit & 0x00000000FFFFFFFF;
 	///int h_moves = (bit & 0xFFFFFFFF00000000) >> 32;
 
-	int count = _mm_popcnt_u64(bit);
+	INT32 count = (INT32)_mm_popcnt_u64(bit);
 
 	return count;
 }
